% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/apcMatrix.R
\name{apcMatrix}
\alias{apcMatrix}
\title{All pairwise comparisons matrix (Tukey contrasts)}
\usage{
apcMatrix(lfm, lev = NULL)
}
\arguments{
\item{lfm}{a \eqn{k\times p} matrix where each line defines a linear
    function to estimate a lsmean (or any linear function). In
    general, these matrices are obtained using
    \code{doBy::LSmatrix()}.}

\item{lev}{a character vector with dimension equals to the numbers of
    lines of \code{lfm} matrix (\eqn{k}). Default is \code{NULL} and
    the row names of code{lfm} is used. If row names is also
    \code{NULL}, an integer sequence is used to identify the
    comparisons.}
}
\value{
a \eqn{K\times p} matrix with the linear functions that
    define all pairwise contrasts. \eqn{K} is \eqn{k\choose 2}.
}
\description{
This function takes a matrix where each line defines a
    linear function of the parameters to estimate a marginal mean
    (aka least squares means) and return the matrix that define the
    contrasts among these means. All pairwise contrasts are returned
    (aka Tukey contrasts). The matrix with these contrasts can be
    passed to \code{multcomp::glht()} to estimate them or used in
    explicit matricial calculus.
}
\examples{
## A matrix of linear functions corresponding to the cell means
## parametrization.
X <- diag(4)

## If no rownames an integer sequence is used.
rownames(X)
apcMatrix(X)

## With rownames, those are used.
rownames(X) <- letters[nrow(X):1]
apcMatrix(X)

## Passing names by the argument `lev=`.
apcMatrix(X, lev=LETTERS[1:nrow(X)])

## Using the special case with attribute "grid" present in matrices
## returned by doBy::LSmeans() and doBy::LSmatrix().
attr(X, "grid") <- data.frame(n=LETTERS[1:nrow(X)])
apcMatrix(X)
}
\author{
Walmes Zeviani, \email{walmes@ufpr.br}
}
\seealso{
\link[doBy]{LSmatrix}.
}

