% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/biasBox.R
\name{biasBox}
\alias{biasBox}
\title{Box bias measaure for nonlinear regression models}
\usage{
biasBox(nls.obj)
}
\arguments{
\item{nls.obj}{An object of class \code{nls} that has gradient and
hessian attributes. See \link[stats]{deriv3}.}
}
\value{
A list with three named elements:
\itemize{
    \item \code{Absolute_bias} is the absolute bias.
    \item \code{Relative_theta} is the bias in relation to the
pontual parameter estimates, in percentage.
    \item \code{Relative_std.error} is the bias in relation to the
precision of the potual parameter estimates, in this case, its
standard error, in percentage.
}
}
\description{
This function calculates the asymptotic Box bias measure for
nonlinear regression models defined by Box (1971). See the references section.
}
\examples{
library(lattice)

data(Puromycin)

xyplot(rate~conc, groups=state, data=Puromycin)

da <- subset(Puromycin, state=="treated")

##-------------------------------------------
## Model 1: Michaelis-Menten.

model1 <- deriv3(expr=~A*conc/(B+conc),
                 namevec=c("A", "B"),
                 function.arg=function(conc, A, B){ NULL })

m1 <- nls(rate~model1(conc, A, B), data=da,
          start=list(A=200, B=0.05))
coef(m1)
bb1 <- biasBox(m1)

##-------------------------------------------
## Model 2: monomolecular.

model2 <- deriv3(expr=~A*(1-exp(-log(2)*conc/B)),
                 namevec=c("A", "B"),
                 function.arg=function(conc, A, B){ NULL })

m2 <- nls(rate~model2(conc, A, B), data=da,
          start=list(A=200, B=0.05))
coef(m2)
bb2 <- biasBox(m2)

##-------------------------------------------
## Bias side by side.

cbind(do.call(rbind, bb1),
      do.call(rbind, bb2))
}
\author{
Walmes Zeviani, \email{walmes@ufpr.br}
}
\references{
Box, M. J. (1971). Bias in nonlinear estimation. Journal
of Royal Statistical Society. Serie B. Methodological, $33 (2), 171–201.
}

