% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/legTools.R
\docType{data}
\name{bib1}
\alias{bib1}
\title{A balanced incomplete block design of type I}
\format{a \code{data.frame} with 56 records and 4 variables.}
\source{
Pimentel Gomes, F. (2009). Curso de Estatística Experimental
    (15th ed.). Piracicaba, São Paulo: FEALQ. (page 190)
}
\usage{
data(bib1)
}
\description{
This data is under a balanced complete block design
    named type I. There are 7 replications that are groups of 4
    blocks of size 2 in a such a way that each treatment occurs once
    in each replication and 7 times at all. There are 8 treatments,
    28 blocks at all. Treatment occur in pairs once. These treatments
    are in fact cells of a \eqn{2^3} factorial design from combining
    nitrogen (P), phosphorus (P) and potassium (K) fertilizers at two
    levels each.

\itemize{

\item \code{rept} a categorical unordered factor with 7 levels. Each
    \code{rept} has 4 bloks of size 2.

\item \code{N} content of nitrogen in the fertilizer (low/high).

\item \code{P} content of phosphorus in the fertilizer (low/high).

\item \code{K} content of potassium in the fertilizer (low/high).

\item \code{block} a categorical unordered factor with 4 levels in
    each \code{rept}, so 28 at all.

\item \code{treat} a categorical unordered factor with 8 levels, the
    treatments studied.

\item \code{y} some response variable. The book doesn't gave details.

}
}
\examples{
require(lattice)

data(bib1)
str(bib1)

xyplot(y~treat|rept, groups=block, data=bib1, type="b",
       ylab="Y", xlab="Treatment")

xyplot(y~treat, data=bib1, jitter.x=TRUE,
       ylab="Y", xlab="Treatment")

xyplot(y~N|P+K, groups=rept, data=bib1, type="b",
       ylab="Y", xlab="Nitrogen")
}
\keyword{datasets}

