% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/legTools.R
\docType{data}
\name{bib2}
\alias{bib2}
\title{A balanced incomplete block design of type II}
\format{a \code{data.frame} with 42 records and 4 variables.}
\source{
Pimentel Gomes, F. (2009). Curso de Estatística Experimental
    (15th ed.). Piracicaba, São Paulo: FEALQ. (page 192)
}
\usage{
data(bib2)
}
\description{
This data is under a balanced complete block design
    named type II. There are 3 replications that are groups of 7
    blocks of size 2 in a such a way that each treatment occurs twice
    in each replication and 6 times at all. There are 7 treatments
    and 21 blocks at all. Treatments occur in pairs once.

\itemize{

\item \code{rept} a categorical unordered factor with 3 levels. Each
    \code{rept} has 7 bloks of size 2.

\item \code{block} a categorical unordered factor with 8 levels in
    each \code{rept}, so 21 at all.

\item \code{treat} a categorical unordered factor with 7 levels, the
    treatments studied.

\item \code{y} some response variable. The book doesn't gave details.

}
}
\examples{
require(lattice)

data(bib2)
str(bib2)

xyplot(y~treat|rept, groups=block, data=bib2, type="b",
       ylab="Y", xlab="Treatment")

xyplot(y~treat, data=bib2, jitter.x=TRUE,
       ylab="Y", xlab="Treatment")
}
\keyword{datasets}

