% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/legTools.R
\docType{data}
\name{bib3}
\alias{bib3}
\title{A balanced incomplete block design of type III}
\format{a \code{data.frame} with 30 records and 3 variables.}
\source{
Pimentel Gomes, F. (2009). Curso de Estatística Experimental
    (15th ed.). Piracicaba, São Paulo: FEALQ. (page 185)
}
\usage{
data(bib3)
}
\description{
This data is under a balanced complete block design
    named type III. There are 5 treatments and 10 blocks of size 3
    plots. Each treatment is repeated 6 times and they occour
    together (in pairs) 3 times.

\itemize{

\item \code{block} a categorical unordered factor with 10 levels.

\item \code{treat} a categorical unordered factor with 5 levels, the
    treatments studied.

\item \code{y} some response variable. The book doesn't gave details.

}
}
\examples{
require(lattice)

data(bib3)
str(bib3)

xyplot(y~treat|block, data=bib3,
       ylab="Y",
       xlab="Treatment")

g <- nlevels(bib3$treat)
a <- seq(0, by=(2*pi)/(g), length.out=g)
y <- sin(a)
x <- cos(a)
plot(y~x, asp=1, xlim=c(-1,1), ylim=c(-1,1))

for (b in levels(bib3$block)){
    cbn <- combn(x=as.integer(bib3$treat[bib3$block==b]),
                 m=2)
    segments(
        x0=x[cbn[1,]], y0=y[cbn[1,]],
        x1=x[cbn[2,]], y1=y[cbn[2,]], col=b)
}
}
\keyword{datasets}

