% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/legTools.R
\docType{data}
\name{bib3asin}
\alias{bib3asin}
\title{A balanced incomplete block design of type III}
\format{a \code{data.frame} with 52 records and 3 variables.}
\source{
Pimentel Gomes, F. (2009). Curso de Estatística Experimental
    (15th ed.). Piracicaba, São Paulo: FEALQ. (page 198)

Fraga Jr., C. G.; Costa, A. S. (1950). Análise de um experimento para
    combate de vira-cabeça do tomateiro. Bragantia, 10:305-316.
}
\usage{
data(bib3asin)
}
\description{
This data is under a balanced complete block design
    named type III. There are 13 treatments and 13 blocks of size 4
    plots. Each treatment is repeated 6 times and they occour
    together (in pairs) only once.

\itemize{

\item \code{block} a categorical unordered factor with 13 levels.

\item \code{treat} a categorical unordered factor with 13 levels, the
    treatments studied.

\item \code{z} correpond a transformation of the original recorded
    variable, the observed percentual of healthy plants, \eqn{p}. So,
    \eqn{z = \arcsin{\sqrt{p/100}}}, in radians is applied to
    stabilize the variance to comply with the assumption of constant
    variance.

}
}
\examples{
require(lattice)

data(bib3asin)
str(bib3asin)

xyplot(z~treat|block, data=bib3asin,
       ylab="Arc sin of heathy plants fraction",
       xlab="Treatment")

## Why not consider a beta distribution for p?
bib3asin$p <- sin(bib3asin$z*pi/180)^2

xyplot(p~treat|block, data=bib3asin,
       ylab="Fraction of healthy plants",
       xlab="Treatment")

g <- nlevels(bib3asin$treat)
a <- seq(0, by=(2*pi)/(g), length.out=g)
y <- sin(a)
x <- cos(a)
plot(y~x, asp=1, xlim=c(-1,1), ylim=c(-1,1))

for (b in levels(bib3asin$block)){
    cbn <- combn(x=as.integer(bib3asin$treat[bib3asin$block==b]),
                 m=2)
    segments(
        x0=x[cbn[1,]], y0=y[cbn[1,]],
        x1=x[cbn[2,]], y1=y[cbn[2,]], col=b)
}
}
\keyword{datasets}

