% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/legTools.R
\docType{data}
\name{cassavaYield}
\alias{cassavaYield}
\title{Cassava variety competition experiment}
\format{a \code{data.frame} with 24 records and 3 variables.}
\source{
Pimentel Gomes, F. (2009). Curso de Estatística Experimental
    (15th ed.). Piracicaba, São Paulo: FEALQ. (page 93)
}
\usage{
data(cassavaYield)
}
\description{
These data are from an experiment done by The Brazilian
    Agricultural Research Corporation in Cassava & Tropical Fruits
    (Centro Nacional de Pesquisa da Mandioca e Fruticultura da
    Embrapa) at Cruz das Almas, Bahia. The experiment was done in a
    randomized complete block design with 4 blocks. Cassava yield
    (t/ha) was recorded in each experimental unit.

\itemize{
    \item \code{block} a categorical unordered factor with 4 levels.
    \item \code{variety} a categorical unordered factor with 6
    levels.
    \item \code{yield} cassava yield (t/ha).
}
}
\examples{
library(lattice)
data(cassavaYield)

plot(yield~variety, data=cassavaYield,
     groups=block, type="o",
     ylab=expression(Yield~(t~ha^{-1})),
     xlab="Variety")
}
\keyword{datasets}

