% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/legTools.R
\docType{data}
\name{castorbeansYield}
\alias{castorbeansYield}
\title{Castor beans variety competition experiments in some locations}
\format{a \code{data.frame} with 45 records and 3 variables.}
\source{
Pimentel Gomes, F. (2009). Curso de Estatística Experimental
    (15th ed.). Piracicaba, São Paulo: FEALQ. (page 149)

Souza, O. Ferreira de.; Canecchio, F. V. (1952). Melhoramento de
    mamoeira, VII. Bragantia 12:301-307.
}
\usage{
data(castorbeansYield)
}
\description{
These data are from a set of experiments evaluating
    varieties of castor beans in terms of yield (kg/ha) for some
    locations (counties).

\itemize{

\item \code{variety} a categorical unordered factor with 8 levels,
    varieties and lines of castor beans.

\item \code{loc} a categorical unordered factor with 5 levels, the
    locations (counties) experimental stations.

\item \code{meanYield} is the mean of yield for a variety in each
    location. So, this the mean across all plots of the same variety
    in each experiment.

}
}
\details{
The data in the book was not complete because doesn't report
    individual plot values but the mean for a variety in each single
    experiment. Neither mention which experimental design was used in
    each station. The book report the Mean Square Error estimates for
    each experiment. These values as provided as an attribute of the
    object, \code{attr(peanut, "MSE")} and they comes from the ANOVA
    table corresponding to an appropriate model for each
    location. With these MSE is possible use them in a such a way
    that a partial ANOVA table can be obtained to test the effect of
    location, variety and its interaction.
}
\examples{
require(lattice)

data(castorbeansYield)
str(castorbeansYield)

xyplot(meanYield~variety, data=castorbeansYield,
       groups=loc, type="o",
       ylab=expression(Yield~(t~ha^{-1})),
       xlab="Variety")
}
\keyword{datasets}

