% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/legTools.R
\docType{data}
\name{cornYield}
\alias{cornYield}
\title{Corn yield as function of fertilization with NPK}
\format{a \code{data.frame} with 32 records and 4 variables.}
\source{
Frederico, P. (2009). Curso de Estatística Experimental (15th
    ed.). Piracicaba, São Paulo: FEALQ. (page 115)
}
\usage{
data(cornYield)
}
\description{
These data are from an \eqn{2^3} factorial experiment
    studing the effect of Nitrogen (N), Phosporus (P) and Potassium
    (K) on corn yield in a randomized block design.

\itemize{
  \item \code{block} a factor with 4 levels.
  \item \code{N} low (-1) and high (+1) levels of nitrogen.
  \item \code{P} low (-1) and high (+1) levels of phosporus.
  \item \code{K} low (-1) and high (+1) levels of potassium.
  \item \code{yield} corn yield (ton/ha).
}
}
\examples{
library(lattice)
library(latticeExtra)

data(cornYield)
str(cornYield)

xyplot(yield~N|P, groups=K,
       data=cornYield, type=c("p", "a"),
       ylab=expression(Yield~(ton~ha^{-1})),
       xlab="Nutrient level")

xyplot(yield~N, groups=interaction(P, K),
       data=cornYield, type=c("p", "a"),
       auto.key=list(columns=2),
       ylab=expression(Yield~(ton~ha^{-1})),
       xlab="Nutrient level")
}
\keyword{datasets}

