% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/legTools.R
\docType{data}
\name{cornYield3}
\alias{cornYield3}
\title{An triple lattice design in corn yield}
\format{a \code{data.frame} with 48 records and 4 variables.}
\source{
Pimentel Gomes, F. (2009). Curso de Estatística Experimental
    (15th ed.). Piracicaba, São Paulo: FEALQ. (page 215)
}
\usage{
data(cornYield3)
}
\description{
This is a experiment evaluating the yield (kg/plot) for
    16 corn hybrids in a triple lattice design.

\itemize{

\item \code{rept} a categorical unordered factor with 3 levels.

\item \code{block} a categorical unordered factor with 4 levels in
    each \code{rept}, so 12 at all. Each block has size 4.

\item \code{hybrid} a unordered factor with 16 levels, the corn
    hybrids studied.

\item \code{yield} the (continuous) response variable of the
    experiment in kg/plot.

}
}
\examples{
require(lattice)

data(cornYield3)
str(cornYield3)

xyplot(yield~hybrid|rept, groups=block,
       data=cornYield3, type="b",
       ylab=expression(Yield~(kg~plot^{-1})),
       xlab="Hybrid")

xyplot(yield~hybrid, data=cornYield3,
       jitter.x=TRUE, type=c("p", "a"),
       ylab=expression(Yield~(kg~plot^{-1})),
       xlab="Hybrid")

g <- nlevels(cornYield3$hybrid)
a <- seq(0, by=(2*pi)/(g), length.out=g)
y <- sin(a)
x <- cos(a)
plot(y~x, asp=1, xlim=c(-1,1), ylim=c(-1,1))

for (b in levels(cornYield3$block)){
    cbn <- combn(x=as.integer(cornYield3$hybrid[cornYield3$block==b]),
                 m=2)
    segments(
        x0=x[cbn[1,]], y0=y[cbn[1,]],
        x1=x[cbn[2,]], y1=y[cbn[2,]], col=b)
}
}
\keyword{datasets}

