% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/legTools.R
\docType{data}
\name{cottonFert}
\alias{cottonFert}
\title{A set of experiments in different locations studing NK on
    cotton}
\format{a \code{data.frame} with 100 records and 4 variables.}
\source{
Pimentel Gomes, F. (2009). Curso de Estatística Experimental
    (15th ed.). Piracicaba, São Paulo: FEALQ. (page 142)
}
\usage{
data(cottonFert)
}
\description{
These data is a set of experiments carried out in
    different locations studing NK fertilization in cotton. All the 5
    experiments are a complete randomized design with 4 replications
    and 5 levels of fertilization based on N and K levels and a
    control.

\itemize{

\item \code{trt} unordered factor, treatment that consist of 4 cells
    from a 2^2 factorial design (\eqn{N\times K}) and a control.

\item \code{rept} integer, indexes experimental units.

\item \code{loc} an unordered factor representing the locations where
    the experiment was carried out.

\item \code{y} numeric, the response variable of the experiment. The
    text book didn't give details.

}
}
\examples{
library(lattice)

data(cottonFert)
str(cottonFert)

xyplot(y~trt|loc,
       data=cottonFert, type=c("p", "a"),
       ylab="y", xlab="Treatment")

xyplot(log(y)~trt|loc,
       data=cottonFert, type=c("p", "a"),
       ylab="y", xlab="Treatment")
}
\keyword{datasets}

