% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/legTools.R
\docType{data}
\name{cowmilkYield3}
\alias{cowmilkYield3}
\title{Milk yield in a switch back experiment}
\format{a \code{data.frame} with 36 records and 5 variables.}
\source{
Pimentel Gomes, F. (2009). Curso de Estatística Experimental
    (15th ed.). Piracicaba, São Paulo: FEALQ. (page 278)

Lucas, H. L. (1956). Switch-back trials for more than two
    treatments. Journal of Dairy Sciences, 39:146-154.
}
\usage{
data(cowmilkYield3)
}
\description{
Data adapted from Lucas (1956). Were studied the effect
    of three treatments on the milk yield of dairy cattle. Twelve
    cows were the experimental units grouped in three blocks of
    different size. Each cow received two treatments, one in the
    first and third periods and the second at the secont period. This
    is a switch back design.

\itemize{

\item \code{period} an ordered categorical factor with 3 levels.

\item \code{block} an nominal factor with 3 levels.

\item \code{cow} a nominal variable that represent each cow of the
    experiment.

\item \code{treat} a nominal categorical factor with 3 levels.

\item \code{yield} the milk yield.

}
}
\examples{
library(lattice)

data(cowmilkYield3)
str(cowmilkYield3)

xyplot(yield~period|cow, groups=treat,
       data=cowmilkYield3, type="o",
       ylab="Milk yield",
       xlab="Period")
}
\keyword{datasets}

