% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/knit2wpCrayon.R
\name{knit2wpCrayon}
\alias{knit2wpCrayon}
\title{knit to wordpress that uses crayon sintax highlight}
\usage{
knit2wpCrayon(input, title = "A post from knitr", ..., action = c("newPost",
  "editPost", "newPage"), postid, encoding = getOption("encoding"),
  upload = FALSE, publish = FALSE, write = TRUE)
}
\arguments{
\item{input}{a markdown or Rmarkdown file.}

\item{title}{title for the post.}

\item{...}{other meta information of the post. See}

\item{action}{indicates a new post, a edition of the post or a new
page.}

\item{postid}{number of the post.}

\item{encoding}{ht encoding of the input file.}

\item{upload}{logical, if the file is to be updated to the blog.}

\item{publish}{logical, if the post is to be published or stay in
draft mode.}

\item{write}{logical, is the result of knit be written to a html
    file. This is useful to copy from this file and paste inside the
    wordpress post editor (on text mode, not visual mode).}
}
\value{
None is returned by the function.
}
\description{
This function improves the \code{RWordPress::knit2wp} to
    allow properly render code when using Crayon Sintax Highlighter.
}
\examples{
\donttest{

library(knitr)
library(RWordPress)

post <- "2015-08-24_polyGui.Rmd"
title <- "Interface para regressão polinomial"
categ <- c("gui", "rbloggers_pt")
keywd <- c("gWdigets", "legTools", "lm", "poly")

pass <- scan(n=1, what=character())
options(WordpressLogin=c(walmes=pass),
        WordpressURL="http://blog.leg.ufpr.br/xmlrpc.php")

knit2wpCrayon(post, title=title,
              action="editPost", postid=179,
              categories=categ, mt_keywords=keywd,
              ## write=TRUE, upload=FALSE,
              write=FALSE, upload=TRUE,
              publish=FALSE)

}
}
\author{
Walmes Zeviani, \email{walmes@ufpr.br}
}

