% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/legTools.R
\docType{data}
\name{kornYield2}
\alias{kornYield2}
\title{Axial factorial NPK experiment with added treatments}
\format{a \code{data.frame} with 9 records and 5 variables.}
\source{
Frederico, P. (2009). Curso de Estatística Experimental (15th
    ed.). Piracicaba, São Paulo: FEALQ. (page 132)

Simão, S. (1960). Estudo da planta e dos frutos da mangueira
    (\emph{Manginifera indica} L.). Piracicaba, 1960. Thesis.
}
\usage{
data(sugarcaneYield4)
}
\description{
These data are from an axial 3 factorial experiment
    studing NPK in the yield of korn. Tow controls were added, one is
    zer control (no NPK) and the other is central factorial point
    plus presence of limestone.

\itemize{
  \item \code{N} content of nitrogen in the fertilizer.
  \item \code{P} content of phosphorus in the fertilizer.
  \item \code{K} content of potassium in the fertilizer.
  \item \code{limestone} presence (1) or absence of limestone (0).
  \item \code{acid} mean of korn yield in 16 locations (ton/ha).
}
}
\details{
The experiment was caried out in 16 different locations but
    only the mean by cell combinations were available in the text
    book.
}
\examples{
library(lattice)
library(latticeExtra)

data(kornYield2)
str(kornYield2)

## Axial triple factorial with 2 controls.
ftable(xtabs(~N+P+K, data=kornYield2))

xyplot(yield~N+P+K,
       groups=as.integer(limestone==1 | (N+P+K)==0),
       data=kornYield2, type=c("p", "a"),
       auto.key=TRUE,
       ylab="Acidity",
       xlab="Month")
}
\keyword{datasets}

