% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/legTools.R
\docType{package}
\name{legTools}
\alias{legTools}
\alias{legTools-package}
\title{Convenience Functions, Small GUI to Teach Statistics and Some
    Datasets.}
\description{
legTools is a collection of R functions and datasets
}

