% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/legTools.R
\docType{data}
\name{mangoAcidity}
\alias{mangoAcidity}
\title{Acidity of mango fruits by varieties, years and months}
\format{a \code{data.frame} with 54 records and 6 variables.}
\source{
Pimentel Gomes, F. (2009). Curso de Estatística Experimental
    (15th ed.). Piracicaba, São Paulo: FEALQ. (page 132)

Simão, S. (1960). Estudo da planta e dos frutos da mangueira
    (\emph{Manginifera indica} L.). Piracicaba, 1960. Thesis.
}
\usage{
data(mangoAcidity)
}
\description{
These data are from an observational study along 3 years
    where acidity in fruits of 6 varieties of mango was determined in
    Novermber, December and January.

\itemize{
  \item \code{variety} a categorical variable with 6 levels that
    represents mango varieties studied.
  \item \code{year} the year of harvesting.
  \item \code{month} the month of harvesting.
  \item \code{acid} mean of the acidity determined in 3 fruits.
}
}
\examples{
library(lattice)

data(mangoAcidity)
str(mangoAcidity)

## reshape::cast() can also be used.
with(mangoAcidity,
     ftable(tapply(acid,
                   list(variety, year, month),
                   FUN=identity)))

xyplot(acid~month|variety, groups=year,
       data=mangoAcidity, type=c("p", "a"),
       auto.key=TRUE,
       ylab="Acidity",
       xlab="Month")
}
\keyword{datasets}

