% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/legTools.R
\docType{data}
\name{peanutYield}
\alias{peanutYield}
\title{Peanut variety competition experiments in some locations and
    years}
\format{a \code{data.frame} with 36 records and 4 variables.}
\source{
Pimentel Gomes, F. (2009). Curso de Estatística Experimental
    (15th ed.). Piracicaba, São Paulo: FEALQ. (page 150)

Souza, O. Ferreira de.; Abramides, Eduardo. (1952). Ensaios de
    variedades de amendoim. Bragantia 12:349-358.
}
\usage{
data(peanutYield)
}
\description{
These data are from a set of experiments evaluating
    varieties of peanut in terms of yield (kg/ha) for some locations
    and years.

\itemize{

\item \code{variety} a categorical unordered factor with 4 levels,
    peanut varieties.

\item \code{loc} a categorical unordered factor with 3 levels, the
    locations (counties) of the experimental stations.

\item \code{year} a categorical factor, the crop year.

\item \code{meanYield} is the adjusted mean of yield for a variety in
    each location and year.

}
}
\details{
The data in the book was not complete because doesn't report
    individual plot values but the adjusted mean for a variety in
    each single experiment. Neither mention which experimental design
    was used in each station. The book report the Mean Square Error
    estimates for each experiment. These values as provided as an
    attribute of the object, \code{attr(peanut, "MSE")} and they
    comes from the ANOVA table corresponding to an appropriate model
    for each location. With these MSE is possible use them in a such
    a way that a partial ANOVA table can be obtained to test the
    effect of location, variety and its interaction.
}
\examples{
require(lattice)

data(peanutYield)
str(peanutYield)

xyplot(meanYield~variety|year, data=peanutYield,
       groups=loc, type="o",
       ylab=expression(Yield~(t~ha^{-1})),
       xlab="Variety")
}
\keyword{datasets}

