% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/legTools.R
\docType{data}
\name{plowing}
\alias{plowing}
\title{Plowing level on corn yield}
\format{a \code{data.frame} with 24 records and 3 variables.}
\source{
Pimentel Gomes, F. (2009). Curso de Estatística Experimental
    (15th ed.). Piracicaba, São Paulo: FEALQ. (page 91)
}
\usage{
data(plowing)
}
\description{
These data are from an experiment done by the engineer
    Duvilio Ometto to study the effect of plowing level on corn
    yield. It was used 2 levels of plowing: normal (or superficial)
    and deep. The experiment was done in a randomized complete block
    design with 6 blocks. Corn yield (t/ha) was recorded in each
    experimental unit but in this experiment there was 2 experimental
    units for each factor level in each block.

\itemize{

\item \code{block} a categorical unordered factor with 6 levels.

\item \code{plow} a categorical unordered factor with 2 levels.

\item \code{yield} corn yield (kg in 200 m\eqn{^2} of area).

}
}
\examples{
library(lattice)
data(plowing)

xyplot(yield~plow|block, data=plowing, type=c("p", "a"),
       ylab=expression(Yield~(t~ha^{-1})),
       xlab="Plowing level")
}
\keyword{datasets}

