% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/poly.gui.R
\name{poly.gui}
\alias{poly.gui}
\title{Polynomial regression GUI}
\usage{
poly.gui(x, y, data, er = 0.05)
}
\arguments{
\item{x,y}{independent and dependent (numeric) regression variables.}

\item{data}{an optional \code{data.frame}.}

\item{er}{stands for extend range. It is used to extend the plotting
range by a fraction on both sides and directions. Default is
0.05. See \link[grDevices]{extendrange}.}
}
\value{
None is returned by the function.
}
\description{
This function opens a interface to control the
polynomial degree in linear regression. It shows the observed values
and the corresponding fitted curve superimposed with confidence bands
(for the fitted value) and show the residuals plot too. It assumes
that \code{gWidgets} and \code{gWidgetstcltk} packages are available.
}
\examples{
\donttest{

poly.gui(x=area, y=peri, data=rock, er=0.3)
poly.gui(x=speed, y=dist, data=cars, er=0.3)
poly.gui(x=eruptions, y=waiting, data=faithful, er=0.3)

}
}

