% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/legTools.R
\docType{data}
\name{potatoYield}
\alias{potatoYield}
\title{Potato variety competition experiment}
\format{a \code{data.frame} with 32 records and 3 variables.}
\source{
Pimentel Gomes, F. (2009). Curso de Estatística Experimental
    (15th ed.). Piracicaba, São Paulo: FEALQ. (page 76)
}
\usage{
data(potatoYield)
}
\description{
These data are from an experiment done by the engineer
    Oscar A. Garay at Balcare, Argentina. The experiment was done in
    a randomized complete block design with 4 blocks. Potato yield
    (t/ha) was recorded in each experimental unit.

\itemize{
    \item \code{block} a categorical unordered factor with 4 levels.
    \item \code{variety} a categorical unordered factor with 8
    levels.
    \item \code{yield} potato yield (t/ha).
}
}
\examples{
library(lattice)
data(potatoYield)

plot(yield~variety, data=potatoYield,
     groups=block, type="o",
     ylab=expression(Yield~(t~ha^{-1})),
     xlab="Variety")
}
\keyword{datasets}

