% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/legTools.R
\docType{data}
\name{potatoYield2}
\alias{potatoYield2}
\title{Potato variety competition experiments in several locations}
\format{a \code{data.frame} with 56 records and 3 variables. There is
    an attribute named \code{MSE}, a named vector containing the Mean
    Squares Errors estimates for each experiment.}
\source{
Pimentel Gomes, F. (2009). Curso de Estatística Experimental
    (15th ed.). Piracicaba, São Paulo: FEALQ. (page 147)
}
\usage{
data(potatoYield2)
}
\description{
These data are from a set of experiments done by the
    engineer Oscar A. Garay at Balcare, Argentina. These experiments
    were done in a randomized complete block design with 4 blocks and
    at 7 locations on the potato production region at the Buenos
    Aires province.

\itemize{

\item \code{variety} a categorical unordered factor with 8 levels,
    varieties of potato.

\item \code{loc} a categorical unordered factor with 7 levels, the
    locations that represent farms or experimental stations.

\item \code{sumYield} is the sum of yield for a variety in each
    experiment. Then, this sum values across 4 blocks in each
    experiment. To get the mean yield you should divide by 4. Yield
    is t/ha.

}
}
\details{
The data in the book was not complete because doesn't report
    individual plot values but, instead, the sum for a variety in
    each experiment. To do a joint or global analysis, with all
    locations, varieties and blocks, its necessary all individual
    plot values. The book report the Mean Square Error estimates for
    each experiment as an attribute of the object,
    \code{attr(potatoYield2, "MSE")} and they comes from the ANOVA
    table in which the model is \code{~block+variety} for each
    location. The data set \link[legTools]{potatoYield} correspond
    the location 3. With these MSE is possible use them in a such a
    way that a partial ANOVA table can be obtained to test the effect
    of location, variety and its interaction.
}
\examples{
require(lattice)

data(potatoYield2)
str(potatoYield2)

xyplot(sumYield/4~variety, data=potatoYield2,
       groups=loc, type="o",
       ylab=expression(Yield~(t~ha^{-1})),
       xlab="Variety")
}
\keyword{datasets}

