% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/runAllChunks.R
\name{runAllChunks}
\alias{runAllChunks}
\title{Run all chunks in a Rmd file}
\usage{
runAllChunks(Rmd, envir = globalenv())
}
\arguments{
\item{Rmd}{the name of the Rmd file.}

\item{envir}{the environment in which the chunks will be
    evaluated. By default it is the GlobalEnv.}
}
\value{
Objects created in the chunks from the Rmd file.
}
\description{
This function was developed to run all chunks in a knitr
    Rmd (R markdown) file at once. Mainly for exploring and debugging
    purposes.
}
\author{
Fernando Mayer, \email{fernando.mayer@ufpr.br}
}
\references{
This function was based on this
    \href{http://stackoverflow.com/questions/24753969/knitr-run-all-chunks-in-an-rmarkdown-document}{SO
    thread}.
}

