% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/legTools.R
\docType{data}
\name{sugarcaneYield}
\alias{sugarcaneYield}
\title{Sugarcane variety experiment}
\format{a \code{data.frame} with 28 records and 3 variables.}
\source{
Pimentel Gomes, F. (2009). Curso de Estatística Experimental
    (15th ed.). Piracicaba, São Paulo: FEALQ. (page 92)
}
\usage{
data(sugarcaneYield)
}
\description{
These data are from an experiment done by The West São
    Paulo State Sugar Mills Cooperative. The experiment was done in a
    randomized complete block design with 4 blocks. Sugarcane yield
    (kg/plot) was recorded in each experimental unit.

\itemize{
    \item \code{block} a categorical unordered factor with 4 levels.
    \item \code{variety} a categorical unordered factor with 7
    levels.
    \item \code{yield} sugarcane yield (kg/plot).
}
}
\examples{
library(lattice)
data(sugarcaneYield)

plot(yield~variety, data=sugarcaneYield,
     groups=block, type="o",
     ylab=expression(Yield~(kg~plot^{-1})),
     xlab="Variety")
}
\keyword{datasets}

