% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/legTools.R
\docType{data}
\name{sugarcaneYield2}
\alias{sugarcaneYield2}
\title{Sugarcane variety competition experiment}
\format{a \code{data.frame} with 28 records and 3 variables.}
\source{
Pimentel Gomes, F. (2009). Curso de Estatística Experimental
    (15th ed.). Piracicaba, São Paulo: FEALQ. (page 96)
}
\usage{
data(sugarcaneYield2)
}
\description{
These data are from an experiment done in a latin square
    design of size 5. Sugarcane yield (kg/plot) was recorded in each
    experimental unit.

\itemize{

\item \code{row} the rows of the latin square that controls in one
    dimention. A categorical unordered factor with 5 levels.

\item \code{col} the columns of the latin square that controls in one
    dimention perpendicular to the previus. A categorical unordered
    factor with 5 levels.

\item \code{variety} a categorical unordered factor with 5 levels.

\item \code{yield} sugarcane yield (kg/plot).

}
}
\examples{
library(lattice)
library(latticeExtra)

xyplot(yield~variety|col,  groups=row, data=sugarcaneYield2,
       ylab=expression(Yield~(kg~plot^{-1})),
       xlab="Variety")

## display.brewer.all()

levelplot(yield~row+col,
          data=sugarcaneYield2, aspect="iso",
          xlab="Row", ylab="Column",
          main=expression(Yield~(kg~plot^{-1})),
          col.regions=colorRampPalette(
              colors=brewer.pal(n=11, name="Spectral")))+
    layer(with(sugarcaneYield2,
               panel.text(x=row, y=col,
                          label=paste(variety, yield))))

aggregate(yield~row, data=sugarcaneYield2, FUN=mean)
aggregate(yield~col, data=sugarcaneYield2, FUN=mean)
aggregate(yield~variety, data=sugarcaneYield2, FUN=mean)
}
\keyword{datasets}

