% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/legTools.R
\docType{data}
\name{sugarcaneYield3}
\alias{sugarcaneYield3}
\title{Sugarcane yield as function of fertilization strategy}
\format{a \code{data.frame} with 28 records and 3 variables.}
\source{
Pimentel Gomes, F. (2009). Curso de Estatística Experimental
    (15th ed.). Piracicaba, São Paulo: FEALQ. (page 99)
}
\usage{
data(sugarcaneYield3)
}
\description{
These data are from an experiment done in a latin square
    design of size 5. Sugarcane yield (kg/plot) was recorded in each
    experimental unit.

\itemize{
  \item \code{row} the rows of the latin square that controls in
    one dimention. A categorical unordered factor with 6 levels.
  \item \code{col} the columns of the latin square that controls in
    one dimention perpendicular to the previus. A categorical
    unordered factor with 6 levels.
  \item \code{fertil} a categorical unordered factor with 6
    levels that is the fertilization strategy applied. These levels
    are a result of treatment cells in a three incomplete factorial
    arrangrment. See detais for more information.
  \item \code{yield} sugarcane yield (kg/plot).
}
}
\details{
The levels of fertilization are in fact a combination of a
    \eqn{3^2} factorial experiment but not all cells are present, so
    this is a (intentional) incomplete three factorial
    experiment. The factors used were limestone (A: present, a:
    absent), \emph{Crotalaria juncae} (B: present, b: absent) and
    fertilizer (C: present, c: absent). Therefore, the level ABC
    means that all three factors are present. To access the effect of
    each factor and interactions can be applied contrasts.
}
\examples{
library(lattice)
library(latticeExtra)

xyplot(yield~fertil|col,  groups=row, data=sugarcaneYield3,
       ylab=expression(Yield~(kg~plot^{-1})),
       xlab="Fertilization", scales=list(x=list(rot=90)))

## display.brewer.all()

levelplot(yield~row+col,
          data=sugarcaneYield3, aspect="iso",
          xlab="Row", ylab="Column",
          main=expression(Yield~(kg~plot^{-1})),
          col.regions=colorRampPalette(
              colors=brewer.pal(n=11, name="Spectral")))+
    layer(with(sugarcaneYield3,
               panel.text(x=row, y=col,
                          label=sprintf("\%s\\n\%0.2f", fertil, yield))))

aggregate(yield~row, data=sugarcaneYield3, FUN=mean)
aggregate(yield~col, data=sugarcaneYield3, FUN=mean)
aggregate(yield~fertil, data=sugarcaneYield3, FUN=mean)

## The incomplete factorial structure.
X <- mapply(FUN=grepl, c("A", "B", "C"),
            MoreArgs=list(x=sugarcaneYield3$fertil))*1
sugarcaneYield3 <- cbind(sugarcaneYield3, as.data.frame(X))

ftable(with(sugarcaneYield3, tapply(yield, list(B, A, C), FUN=mean)))
aggregate(yield~A+B+C, data=sugarcaneYield3, FUN=mean)
}
\keyword{datasets}

