% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/legTools.R
\docType{data}
\name{sugarcaneYield4}
\alias{sugarcaneYield4}
\title{Triple factorial NPK fertilization on sugar cane yield}
\format{a \code{data.frame} with 54 records and 6 variables.}
\source{
Pimentel Gomes, F. (2009). Curso de Estatística Experimental
    (15th ed.). Piracicaba, São Paulo: FEALQ. (page 126)

Straus, E. (1951). Experimentos de adubação na zona canavieira de
    Pernambuco. In: Anais da III Reunião Brasileira de Ciência do
    Solo. 1:336-443.
}
\usage{
data(sugarcaneYield4)
}
\description{
These data are from an \eqn{3^3} factorial experiment
    studing the effect of NPK on the yield of sugar cane.

\itemize{

  \item \code{block} a local control factor with 3 levels.

  \item \code{rept} factor with 2 levels.

  \item \code{N} integer coded nitrogen levels (0, 1, 2).

  \item \code{P} integer coded phosphorus levels (0, 1, 2).

  \item \code{K} integer coded potassium levels (0, 1, 2).

  \item \code{yield} sugar cane yield (ton/ha).

}
}
\details{
There is a missprint in the book for the 9th entry, which
    has yield 59.0, that is coded as 202 istead of 220.
}
\examples{
library(lattice)
library(latticeExtra)

data(sugarcaneYield4)
str(sugarcaneYield4)

xyplot(yield~N|P, groups=K,
       auto.key=list(title="Potassim level", columns=3),
       strip=strip.custom(var.name="Phosphorus", strip.names=TRUE,
                          strip.levels=TRUE, sep=": "),
       data=sugarcaneYield4, type=c("p", "a"),
       ylab=expression(Yield~(ton~ha^{-1})),
       xlab="Nitrogen level level")
}
\keyword{datasets}

