% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/legTools.R
\docType{data}
\name{vinasseFert}
\alias{vinasseFert}
\title{Fertilization with vinasse and mineral}
\format{a \code{data.frame} with 16 records and 4 variables.}
\source{
Frederico, P. (2009). Curso de Estatística Experimental (15th
    ed.). Piracicaba, São Paulo: FEALQ. (page 119)
}
\usage{
data(vinasseFert)
}
\description{
These data are from an \eqn{2^2} factorial experiment
    studing the effect of fertilizaton with vinasse (organic font)
    and complete mineral fertilization.

\itemize{
  \item \code{block} a factor with 4 levels.
  \item \code{mineral} low (-1) and high (+1) levels of mineral
    fertilization.
  \item \code{vinasse} low (-1) and high (+1) levels of fetilization
    with vinasse.
  \item \code{y} some response variable. The text book doesn't give
    any information.
}
}
\examples{
library(lattice)
library(latticeExtra)

data(vinasseFert)
str(vinasseFert)

xyplot(y~vinasse, groups=mineral,
       auto.key=list(title="Mineral", columns=2),
       data=vinasseFert, type=c("p", "a"),
       ylab="y",
       xlab="Vinasse level")

m0 <- lm(y~block+(vinasse+mineral)^2, data=vinasseFert)
par(mfrow=c(2,2)); plot(m0); layout(1)
anova(m0)

m1 <- update(m0, .~block+vinasse)
par(mfrow=c(2,2)); plot(m1); layout(1)

anova(m0, m1)
anova(m1)

summary(m1)
}
\keyword{datasets}

