% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/legTools.R
\docType{data}
\name{wgCattle}
\alias{wgCattle}
\title{Grazing intensity and \emph{Brachiaria} species in the weight
    o cattle.}
\format{a \code{data.frame} with 48 records and 5 variables.}
\source{
Pimentel Gomes, F. (2009). Curso de Estatística Experimental
    (15th ed.). Piracicaba, São Paulo: FEALQ. (page 278)

Pimentel Gomes, F.; Nunes, S. G.; Gomes, S. de B.; Curvo,
    J. B. E. (1988). Modificação da análise de variância de ensaios
    de pastejo com bovinos, considerando os blocos de
    animais. Pesquisa Agrpecuária Brasileira 23(9):251-276.
}
\usage{
data(wgCattle)
}
\description{
Data adapted from Pimentel Gomes (1988). The experiment
    is a 2 grazing intensity x 3 species of \emph{Brachiaria}
    factorial. The experimental units were 48 animal grouped in 8
    groups of 6 animals. The 8 groups were splited in two blocks. The
    response variable recorded were the weight of the animals.

\itemize{

\item \code{block} an categorical factor with 2 levels.

\item \code{group} an categorial factor with 4 levels into each
    block.

\item \code{specie} a categorical factor that is species of
    \emph{Brachiaria} for pasture.

\item \code{inten} a numeric factor with 2 levels that is the number
    of grazing animal per unit area.

\item \code{weight} the recorded weight.

}
}
\examples{
library(lattice)

data(wgCattle)
str(wgCattle)

xyplot(weight~specie|group, groups=inten,
       data=wgCattle, type=c("p", "a"),
       auto.key=list(columns=2, title="Intensity (un/ha)"),
       scales=list(x=list(font=3)),
       ylab="Weight (kg)",
       xlab=expression(Species~of~italic("Brachiaria")))
}
\keyword{datasets}

