% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/legTools.R
\docType{data}
\name{wgChickens}
\alias{wgChickens}
\title{Weight gain in chickens as function of sorghum in the feed}
\format{a \code{data.frame} with 16 records and 4 variables.}
\source{
Pimentel Gomes, F. (2009). Curso de Estatística Experimental
    (15th ed.). Piracicaba, São Paulo: FEALQ. (page 267)

Torres A. P.; Pimentel Gomes, F. (1958). Substituição de subprodutos
    de trigo pelo sorgo moído na alimentação de pintos. Escola
    Superior de Agricultura "Luiz de Queiroz". Anais
    E.S.A. "Luiz de Queiroz" 16: 251-276.
}
\usage{
data(wgChickens)
}
\description{
This experiment considered four levels of sorghum in the
    feed and chickens of both sex, so it is a \eqn{4 x 2} factorial
    design and was performed in a randomized complete design with two
    replications. Each experimental unit is a set of 13 animals in
    which the total weight is recorded (decagrams). At the end of the
    experiment, 4 week later, some animals died resulting in some
    unbalancing.

\itemize{

\item \code{gender} a factor with two levels represengin the gerder
    of the chickens.

\item \code{sorghum} a numeric variable at 4 levels that is the
    concentration, in percentage, of sorghum in the feed.

\item \code{n} the number of animals in each experital unit. All
    experimental units had initiated with 13 animals but some of them
    died remaning 12 animals for some experimental units.

\item \code{tw} is the total wieght of each experimental unit in
    decagrams.

}
}
\examples{
library(lattice)

data(wgChickens)
str(wgChickens)

xyplot(tw/n~sorghum, groups=gender,
       data=wgChickens, type=c("p", "a"),
       auto.key=list(
           columns=2, corner=c(0.95, 0.95), title="Gender"),
       ylab="Mean weight gain (kg)",
       xlab="Sorghum concentration in the feed (\%)")
}
\keyword{datasets}

