% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/legTools.R
\docType{data}
\name{wgpigs}
\alias{wgpigs}
\title{Feeding type in pig weight gain}
\format{a \code{data.frame} with 20 records and 2 variables.}
\source{
Frederico, P. (2009). Curso de Estatística Experimental (15th
    ed.). Piracicaba, São Paulo: FEALQ. (page 62)
}
\usage{
data(wgpigs)
}
\description{
This is an artifial dataset corresponding a experiment
    to study the effect of feeding type (factor with 4 categorical
    nominal levels) in pig weight gain. The experiment was a
    randomized complete design with five experimental units per
    treatment level. The experimental unit was a pig. The response
    measured was weight gain from the beggining to the end of the
    experiment.

\itemize{
    \item \code{ft} feeding type, a categorical factor with 4
    levels.
    \item \code{wg} weight gain (kg).
}
}
\examples{
require(lattice)
data(wgpigs)

xyplot(wg~ft, data=wgpigs,
       ylab="Weight gain (kg)",
       xlab="Feeding type")
}
\keyword{datasets}

