% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/yscale.components.right.R
\name{yscale.components.right}
\alias{yscale.component.right}
\alias{yscale.components.right}
\title{y-axis annotations on the right side}
\source{
When such feature was necessary, a search in the web was done
    and a post in the r-help mailing list inspired us
    \code{http://r.789695.n4.nabble.com/Spacing-between-lattice-panels-td855613.html}.
}
\usage{
yscale.component.right(...)
}
\arguments{
\item{...}{arguments passed by the lattice function called. See
    \link[lattice]{yscale.components.default}.}
}
\description{
This function if for place y axis annotation on the
    right side of the plot.
}
\examples{
library(lattice)
library(latticeExtra)

## alternating=2 works when relation="same".
p1 <- xyplot(yield~K|N+P, data=npk,
             scales=list(y=list(alternating=2)))
useOuterStrips(p1)

## y annotation don't is written on the right side.
p2 <- xyplot(yield~K|N+P, data=npk,
             scales=list(y=list(relation="free", alternating=2)))
useOuterStrips(p2)

## The desired result.
p3 <- xyplot(yield~K|N+P, data=npk,
             scales=list(y=list(relation="free", alternating=2)),
             ylab=NULL, ylab.right="Yield",
             yscale.component=yscale.component.right,
             between=list(x=0.5, y=0.2),
             par.settings=list(
                 layout.widths=list(
                     right.padding=-2,
                     left.padding=-2,
                     ylab.right=5),
                 strip.background=list(col=c("gray50", "gray90"))),
             )
useOuterStrips(p3)
}
\author{
Walmes Zeviani, \email{walmes@ufpr.br}
}

