% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/legTools.R
\docType{data}
\name{filterCake}
\alias{filterCake}
\title{Fertilization with filter cake and mineral}
\format{a \code{data.frame} with 16 records and 4 variables.}
\source{
Frederico, P. (2009). Curso de Estatística Experimental (15th
    ed.). Piracicaba, São Paulo: FEALQ. (page 120)
}
\usage{
data(filterCake)
}
\description{
These data are from an \eqn{2^2} factorial experiment
    studing the effect of fertilizaton with filter cake (a residual
    from industrial processing of sugar cane) and traditional mineral
    fertilization.

\itemize{
  \item \code{block} a factor with 4 levels.
  \item \code{mineral} low (-1) and high (+1) levels of mineral
    fertilization.
  \item \code{cake} low (-1) and high (+1) levels of fetilization
    with filter cake.
  \item \code{y} some response variable. The text book doesn't give
    any information.
}
}
\examples{
library(lattice)
library(latticeExtra)

data(filterCake)
str(filterCake)

xyplot(y~cake, groups=mineral,
       auto.key=list(title="Mineral", columns=2),
       data=filterCake, type=c("p", "a"),
       ylab="y",
       xlab="Filter cake level")
}
\keyword{datasets}

