% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/legTools.R
\docType{data}
\name{wgPigs2}
\alias{wgPigs2}
\title{Age of castration in pig weight gain}
\format{a \code{data.frame} with 16 records and 4 variables.}
\source{
Pimentel Gomes, F. (2009). Curso de Estatística Experimental
    (15th ed.). Piracicaba, São Paulo: FEALQ. (page 110)
}
\usage{
data(wgPigs2)
}
\description{
This is an artifial dataset corresponding a experiment
    to study the effect of feeding type (factor with 4 categorical
    nominal levels) in pig weight gain. The experiment was a
    randomized complete design with five experimental units per
    treatment level. The experimental unit was a pig. The response
    measured was weight gain from the beggining to the end of the
    experiment.

\itemize{
  \item \code{litter} a categorical factor with 4 levels that
    represents the rows of the lattin square design and control for
    the differences among litters.
  \item code{size} a categorical ordered variable that represents the
    columns of latin square desing and control for the weight of the
    animals at the beggining of the experiment.
  \item \code{age} age of the animal (days) when castration was
    done. \code{controls} are the animals without castration.
  \item \code{wg} weight gain (kg) after 252 days.
}
}
\examples{
library(lattice)

data(wgPigs2)
str(wgPigs2)

xyplot(wg~age, data=wgPigs2, groups=litter,
       ylab="Weight gain (kg)",
       xlab="Age at castration (days)")
}
\keyword{datasets}

